src/mylib/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio
numpy
Pillow